cloud_function/requirements.txt
requests==2.21.0
google-auth==1.18.0
google-cloud-storage>=1.10.0,<2